+++ Changelog.md
@@ -10,6 +10,7 @@ Types of changes: New, Improved, Fixed, Removed.
 ## [Unreleased]
 
 - Fixed: Issue where some invalid embellishment errors would be missing the file location
+- Fixed: Zed issue with path quoting
 
 ## [1.0.0]