+++ Changelog.md
@@ -11,6 +11,7 @@ Types of changes: New, Improved, Fixed, Removed.
 
 - Fixed: Issue where some invalid embellishment errors would be missing the file location
 - Fixed: Zed issue with path quoting
+- Fixed: Single-note variations at the end of a bar would close after the barline 
 
 ## [1.0.0]